Add first backend route: /api/todos
Add first backend route: /api/todos. Now there is an error when clicking
the Add Todo button because when the code gets to Todo object, it is not
defined (inside our routes.js). So we 'require' it but now we have to
defined our Todo model. NEXT: Define our Todo Model

Contact: joaquin@isla.io