Implement Enhancement #5: Add line-through
Implement Enhancement #5: Add line-through. To add the functionality of
crossing out the text, we use two HTML tags, the first one will show
when the checkbox is not checked, and then the second HTML tag (which is
a copy of the first one) will replace/show the same text that the first
one was showing but with line-through style.

Contact: joaquin@isla.io